HTCONDOR-1768 Remove some most casts of time_t to int

+++ src/condor_collector.V6/collector_stats.cpp
@@ -516,13 +516,13 @@ void UpdatesStats::SetWindowSize(int window)
 
 void UpdatesStats::Publish(ClassAd & ad, int flags) const
 {
-	ad.Assign("StatsLifetime", (int)StatsLifetime);
-	ad.Assign("StatsLastUpdateTime", (int)StatsLastUpdateTime);
+	ad.Assign("StatsLifetime", StatsLifetime);
+	ad.Assign("StatsLastUpdateTime", StatsLastUpdateTime);
 	if (flags & IF_RECENTPUB) {
-		ad.Assign("RecentStatsLifetime", (int)RecentStatsLifetime);
+		ad.Assign("RecentStatsLifetime", RecentStatsLifetime);
 		if (flags & IF_VERBOSEPUB) {
-			ad.Assign("RecentWindowMax", (int)RecentWindowMax);
-			ad.Assign("RecentStatsTickTime", (int)RecentStatsTickTime);
+			ad.Assign("RecentWindowMax", RecentWindowMax);
+			ad.Assign("RecentStatsTickTime", RecentStatsTickTime);
 		}
 	}
 	Pool.Publish(ad, flags);
