+++ src/condor_starter.V6.1/vanilla_proc.cpp
@@ -1010,7 +1010,7 @@ VanillaProc::outOfMemoryEvent() {
 	// lower than the limit when the OOM killer fired.
 	// So have some slop, just in case.
 	if (usageMB < (0.9 * (m_memory_limit / (1024 * 1024)))) {
-		dprintf(D_ALWAYS, "Evicting job because system is out of memory, even though the job is below requested memory: Usage is %ld Mb limit is %ld\n", usageMB, m_memory_limit);
+		dprintf(D_ALWAYS, "Evicting job because system is out of memory, even though the job is below requested memory: Usage is %lld Mb limit is %lld\n", (long long)usageMB, (long long)m_memory_limit);
 		Starter->jic->notifyStarterError("Worker node is out of memory", true, 0, 0);
 		Starter->jic->allJobsGone(); // and exit to clean up more memory
 		return 0;
